[move] TenantRegisterControllerをAuthディレクトリに移動し、ルートを修正
`TenantRegisterController`を`Auth`ディレクトリに移動し、ルートファイルでの参照を修正しました。これにより、認証関連のコントローラが統一され、構造が整理されました。
